Move IdentityEnsembleArray to networks.
